separate command into their own files
